Crate vulkano_blas

vulkano_blas is a library of GLSL Vulkano shaders for BLAS operations.

Crate status

Currently due to the limitations of Vulkano very few BLAS operations can be implemented (my issue on GitHub concerning this).

All shaders I am certain I will implement but cannot implement at the moment due to this limitation are marked as deprecated.

I hope perhaps this crate while at the moment extremely limited might spike some interest and motiviation in the topic.

A note on testing

None of the rustdoc code is tested due to the extreme awkwardness of this. A test is run for every (implemented) shader through, just not in the rustdoc comments.

Modules

isamax	Deprecated
Gets index of max value in vector.

sasum	Deprecated
Gets sum a vector.

saxpy
Constant times a vector plus a vector.

sdot	Deprecated
Vector dot product.

sgemm	Deprecated
Matrix-matrix multiply

sgemv	Deprecated
Matrix-vector multiply.

sscal
Scales a vector by a constant.