Module vulkano_blas::sscal

Scales a vector by a constant.

$ x = a * x $

Where x is vector and a is a scalar.

http://www.netlib.org/lapack/explore-html/d9/d04/sscal_8f.html

// Sets the data buffer (x vector).
// CLoning x here is unneccessary for functionality, it is only convenient such that we can use it later for checking the result.
let x = 0..5u32;
let data_buffer_x = CpuAccessibleBuffer::from_iter(device.clone(), BufferUsage::all(), false,x.clone()).expect("failed to create buffer");

// Initiates a scalar.
let a = 2u32;

// Gets the layout of the pipeline (x,y,z)
let layout = compute_pipeline.layout().descriptor_set_layout(0).unwrap();
// Descriptors push buffers to the compute pipeline.
// Intiates descriptor set.
let set = Arc::new(PersistentDescriptorSet::start(layout.clone())
    .add_buffer(data_buffer_x.clone()).unwrap()
    .build().unwrap()
);

// Intiates command buffer for compute pipeline (1 work groups)
let mut builder = AutoCommandBufferBuilder::new(device.clone(), queue.family()).unwrap();
// Dispatches command buffer builder passing a scalar as push constant.
builder.dispatch([1024,1,1],compute_pipeline.clone(),set.clone(),a).unwrap();

let command_buffer = builder.build().unwrap();

// Submit command buffer.
let finished = command_buffer.execute(queue.clone()).unwrap();

// Halts thread until command buffer executed.
finished.then_signal_fence_and_flush().unwrap()
.wait(None).unwrap();

// Checks content of data buffer.
let x_content = data_buffer_x.read().unwrap();
for (new_x, x) in x_content.iter().zip(x) {
    assert_eq!(*new_x, x * a);
}

Modules

ty

Structs

Layout

MainInput

MainInputIter

MainOutput

MainOutputIter

Shader

SpecializationConstants